requirements.txt
# Core
torch==1.13.1
torchvision==0.14.1
numpy
Pillow
matplotlib

# Nearest neighbor search
annoy

# OpenCLIP
git+https://github.com/mlfoundations/open_clip.git
